# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WyAPI-Golang",
    "version": "2.0.0",
    "description": "网易云音乐解析与下载 API"
  },
  "servers": [
    { "url": "/" }
  ],
  "components": {
    "securitySchemes": {
      "ApiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "status": { "type": "integer" },
          "success": { "type": "boolean" },
          "msg": { "type": "string" },
          "message": { "type": "string" },
          "data": { "type": "object" }
        }
      }
    }
  },
  "paths": {
...
    "/api/info": {
      "get": {
        "summary": "API信息",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/api/music/url": {
      "post": {
        "summary": "获取歌曲播放链接",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "level": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/api/music/detail": {
      "post": {
        "summary": "获取歌曲详情",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/api/music/lyric": {
      "post": {
        "summary": "获取歌词",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/api/music/playlist": {
      "post": {
        "summary": "获取歌单详情",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/api/music/album": {
      "post": {
        "summary": "获取专辑详情",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "搜索音乐",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keyword": { "type": "string" },
                  "limit": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/download": {
      "post": {
        "summary": "下载音乐",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "quality": { "type": "string" },
                  "format": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok" }
        }
      }
    }
  }
}